check customer exists before you delete

+++ src/main/java/com/singh/springdemo/controller/CustomerRestController.java
@@ -19,15 +19,15 @@ import com.singh.springdemo.service.CustomerService;
 @RestController
 @RequestMapping("/api")
 public class CustomerRestController {
-	
+
 	@Autowired
 	private CustomerService customerService;
-	
+
 	@GetMapping("/customers")
 	public List<Customer> getCustomers() {
 		return customerService.getCustomers();
 	}
-	
+
 	@GetMapping("/customers/{customerId}")
 	public Customer getCustomer(@PathVariable int customerId) {
 		Customer customer = customerService.getCustomer(customerId);
@@ -36,7 +36,7 @@ public class CustomerRestController {
 		}
 		return customer;
 	}
-	
+
 	// add mapping for POST / customers - add new customer
 	@PostMapping("/customers")
 	public Customer addCustomer(@RequestBody Customer customer) {
@@ -46,16 +46,21 @@ public class CustomerRestController {
 		customerService.saveCustomer(customer);
 		return customer;
 	}
-	
+
 	@PutMapping("/customers")
 	public Customer updateCustomer(@RequestBody Customer customer) {
 		customerService.saveCustomer(customer);
 		return customer;
 	}
-	
+
 	@DeleteMapping("/customers/{customerId}")
-	public void deleteCustomer(@PathVariable int customerId) {
+	public String deleteCustomer(@PathVariable int customerId) {
+		Customer customer = customerService.getCustomer(customerId);
+		if (customer == null) {
+			throw new CustomerNotFoundException("Customer not found with id: " + customerId);
+		}
 		customerService.deleteCustomer(customerId);
+		return "Delete Customer Id: " + customerId;
 	}
 
 }
